% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcor.R
\name{print.bcor}
\alias{print.bcor}
\title{Print bcor Object}
\usage{
\method{print}{bcor}(x, ...)
}
\arguments{
\item{x}{A bcor object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print bcor Object
}
