% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcor-package.R
\docType{package}
\name{rbcor-package}
\alias{rbcor}
\alias{rbcor-package}
\title{rbcor: Read BCOR Files into R Sparse Matrices}
\description{
Reads binary correlation (BCOR) files generated by LDStore into R's sparse matrix format. BCOR files store linkage disequilibrium correlation matrices in a compressed binary format.
}
\author{
\strong{Maintainer}: Masahiro Kanai \email{mkanai@broadinstitute.org}

}
\keyword{internal}
